requirements.txt
django==3.2.6
djangorestframework==3.12.4
requests==2.26.0
drf-yasg==1.20.0
psycopg2==2.9.1
